requirements.txt

pandas
requests
tqdm
python-dotenv
openai
kaggle
locust
pyarrow
fastparquet